Moved the ByteArrayServletInputStream class to the servlet sub-package as per code review comments.
